Add option to just link original collection
This was already the behavior, so it's enabled by default.
The new behavior is when you disable the option, which makes it just link the original collection and nothing else; transforms are lost.
I might want to make it optional to add the empty in the future as well, not sure what to call it though.
Some of this functionality would be good to add into PowerLink too; perhaps ImperfectLink won't mind.